requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.4.4
chardet==3.0.4
discord.py==1.2.5
entrypoints==0.3
flake8==3.7.9
idna==2.8
mccabe==0.6.1
multidict==4.5.2
mypy==0.750
mypy-extensions==0.4.3
pycodestyle==2.5.0
pyflakes==2.1.1
typed-ast==1.4.0
typing-extensions==3.7.4.1
websockets==6.0
yarl==1.3.0
